refactor(today): extract recovery adaptation logic
